Added Fibonacci built-ins
Added the functions `ﬂ` for the nth Fibonacci number, `ﬁ` that checks if a number is in the fibonacci sequence, `ƒ` that prints the first n  Fibonacci numbers, starting from `0` and `1`. #TODO: Functions for the next and previous Fibonacci numbers, function for the closest fibonacci number, function for all the fibonacci numbers smaller than the given value.